requirements.txt
# requirements.txt
streamlit
requests
jinja2
python-dotenv
google-generativeai
chromadb